Add test framework

+++ Makefile
@@ -1,10 +1,14 @@
 all:
 	gcc -fPIC -g -c -Wall src/*.c
-	mkdir lib
+	mkdir -p lib
 	gcc -shared *.o -o lib/libdecap.so
+tests:
+	gcc -g -o runtests test/*.c
+	./runtests
 
 clean:
+	rm -f runtests
 	rm -f lib/*.so
-	rmdir lib
-	rm -f *.o
+	if test -d lib; then rmdir lib;	fi
 
+	rm -f *.o
